requirements.txt
asgiref==3.11.0
autopep8==2.3.2
backports.zstd==1.3.0
brotli==1.2.0
dj-database-url==3.1.0
Django==5.1.5
django-cors-headers==4.9.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
inflate64==1.0.4
multivolumefile==0.2.3
numpy==2.4.1
openpyxl==3.1.5
pandas==2.3.3
psutil==7.2.1
psycopg==3.3.2
psycopg2-binary==2.9.11
py7zr==1.1.0
pyarrow==22.0.0
pybcj==1.0.7
pycodestyle==2.14.0
pycryptodomex==3.23.0
pyppmd==1.3.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
six==1.17.0
sqlparse==0.5.5
texttable==1.7.0
typing_extensions==4.15.0
tzdata==2025.3
whitenoise==6.11.0
tqdm==4.67.1
